first pass, links fine, 11 tests fail
